+++ libs/ngx-mime/src/lib/core/models/index.ts
@@ -1,3 +1,4 @@
+export { ModeChanges } from './mode-changes';
 export {
   RecognizedTextMode,
   RecognizedTextModeChanges,
