+++ libs/ngx-mime/src/lib/view-dialog/icon/icon.component.scss
@@ -1,9 +1,5 @@
 .mat-icon {
   vertical-align: middle;
-  display: inline-block;
-  fill: currentColor;
-  height: 24px;
-  width: 24px;
 }
 
 .single-page-display {
